Add common code
